restruct : Suppression de GradesScreen et passage sur GradesScreenNew
